requirements.txt
pytest==3.7.2
pytest-selenium==1.13.0
webdriver-manager==1.7
allure-pytest==2.8.6
pytest-bdd==3.2.1
requests==2.20.0